Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.13.0
scikit-learn==1.3.2
xgboost==1.7.6
requests==2.31.0
python-multipart==0.0.6
pickle5==0.0.12
numpy==1.24.3
pandas==2.0.3